+++ client/src/components/login/LoginForm/LoginForm.module.scss
@@ -1,0 +1,5 @@
+.form {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
